Move test and only keep the UI logic part being tested
